[ENH] refactoring by remove user component to app.
