+++ app/Routes/web.php
@@ -5,10 +5,10 @@
  */
 global $app;
 
-use Wepesi\Controller\exampleController;
+use App\controller\exampleController;
+use app\Middleware\exampleValidation;
 use Wepesi\Core\Http\Response;
 use Wepesi\Core\View\View;
-use Wepesi\Middleware\Validation\exampleValidation;
 
 $router = $app->router();
